Allow autoloadBlocks to handle filters or styles without block object
Before a recent change, it was possible to use the autoloadBlocks
functionality to register filters or styles without actually exporting a
block settings object. 2252caa97 broke that functionality by causing an
uncaught type error if a file passed to autoloadBlocks didn't export a
settings object. This adds null checking around the settings object
before accessing it.